requirements.txt
dlib==20.0.0
imgaug==0.4.0
retinaface-pytorch==0.0.7
tensorboard==2.17.0
torch==2.1.2+cu118
torchvision==0.16.2+cu118
transformers==4.42.0
albumentations==2.0.8
opencv-python==4.9.0.80
numpy==1.26.4
pandas==2.3.3
pillow==10.3.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
tqdm==4.64.1
huggingface-hub==0.36.0
safetensors==0.7.0
matplotlib==3.9.0